requirements.txt
streamlit>=1.42.0
langchain>=0.3.17
langchain-community>=0.3.16
langchain-mistralai>=0.2.6
langchain-huggingface>=0.1.2
langchain-text-splitters>=0.3.7
langchain-qdrant>=0.2.0
qdrant-client>=1.13.2
mistralai>=1.5.0
gitpython>=3.1.44
python-dotenv>=1.0.1
sentence-transformers>=3.4.1
duckduckgo-search>=6.3.11
reportlab>=4.2.5
pdfplumber>=0.11.4
numpy>=2.2.3
requests>=2.32.3
SpeechRecognition>=3.12.0
gTTS>=2.5.4
pydub>=0.25.1
